requirements.txt
click==8.1.7
-e git+ssh://git@github.com/k0dik3v/exifedit.git@95309eed5319611179b8f8c16cf710c2fb050577#egg=edit_exif
iniconfig==2.0.0
packaging==24.1
pillow==10.3.0
pluggy==1.5.0
plum-py==0.8.7
pytest==8.2.2
setuptools==70.1.0
wheel==0.43.0
